Test, refactor, and adjust
