+++ frontend/src/routes/app._index/route.tsx
@@ -1,0 +1,86 @@
+import React from "react";
+import { useSelector } from "react-redux";
+import {
+  ClientActionFunctionArgs,
+  json,
+  useLoaderData,
+  useRouteError,
+} from "@remix-run/react";
+import { RootState } from "#/store";
+import AgentState from "#/types/AgentState";
+import FileExplorer from "#/components/file-explorer/FileExplorer";
+import { retrieveFiles, retrieveFileContent } from "#/api/open-hands";
+import { useSocket } from "#/context/socket";
+import { FilesProvider } from "#/context/files";
+import CodeEditorCompoonent from "./code-editor-component";
+
+export const clientLoader = async () => {
+  const token = localStorage.getItem("token");
+  return json({ token });
+};
+
+export const clientAction = async ({ request }: ClientActionFunctionArgs) => {
+  const token = localStorage.getItem("token");
+
+  const formData = await request.formData();
+  const file = formData.get("file")?.toString();
+
+  let selectedFileContent: string | null = null;
+
+  if (file && token) {
+    selectedFileContent = await retrieveFileContent(token, file);
+  }
+
+  return json({ file, selectedFileContent });
+};
+
+export function ErrorBoundary() {
+  const error = useRouteError();
+
+  return (
+    <div className="w-full h-full border border-danger rounded-b-xl flex flex-col items-center justify-center gap-2 bg-red-500/5">
+      <h1 className="text-3xl font-bold">Oops! An error occurred!</h1>
+      {error instanceof Error && <pre>{error.message}</pre>}
+    </div>
+  );
+}
+
+function CodeEditor() {
+  const { token } = useLoaderData<typeof clientLoader>();
+
+  const { runtimeActive } = useSocket();
+  const agentState = useSelector(
+    (state: RootState) => state.agent.curAgentState,
+  );
+
+  const [files, setFiles] = React.useState<string[]>([]);
+
+  React.useEffect(() => {
+    // only retrieve files if connected to WS to prevent requesting before runtime is ready
+    if (runtimeActive && token) retrieveFiles(token).then(setFiles);
+  }, [runtimeActive, token]);
+
+  // Code editing is only allowed when the agent is paused, finished, or awaiting user input (server rules)
+  const isEditingAllowed = React.useMemo(
+    () =>
+      agentState === AgentState.PAUSED ||
+      agentState === AgentState.FINISHED ||
+      agentState === AgentState.AWAITING_USER_INPUT,
+    [agentState],
+  );
+
+  return (
+    <FilesProvider defaultPaths={files}>
+      <div className="flex h-full w-full bg-neutral-900 relative">
+        <FileExplorer />
+        <div className="flex flex-col min-h-0 w-full pt-3">
+          <div className="flex grow items-center justify-center">
+            <CodeEditorCompoonent isReadOnly={!isEditingAllowed} />
+          </div>
+        </div>
+      </div>
+    </FilesProvider>
+  );
+}
+
+export default CodeEditor;
